LayoutTests/media/modern-media-controls/mute-support/mute-support-button-click-expected.txt

Testing the MuteSupport behavior by clicking on the mute button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.muteButton.muted is false

Dispatching click event to the mute button

volumechange event was dispatched
PASS media.muted is true
PASS mediaController.controls.muteButton.muted is true

Dispatching click event to the mute button

volumechange event was dispatched
PASS media.muted is false
PASS mediaController.controls.muteButton.muted is false

PASS successfullyParsed is true

TEST COMPLETE
